Control.lua split to 4 modules Those are remove, build, global and render for now
